requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiosqlite==0.17.0
annotated-types==0.6.0
asyncpg==0.29.0
click==8.1.7
Deprecated==1.2.14
dictdiffer==0.9.0
html5tagger==1.3.0
httptools==0.6.1
importlib-resources==6.1.1
iso8601==1.1.0
limits==3.7.0
multidict==6.0.4
packaging==23.2
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sanic==23.6.0
sanic-ext==23.6.0
sanic-limiter==0.1.3
sanic-routing==23.6.0
six==1.16.0
tomlkit==0.12.3
tortoise-orm==0.20.0
tracerite==1.1.1
typing_extensions==4.9.0
ujson==5.9.0
uvloop==0.19.0
websockets==12.0
wrapt==1.16.0
